refactor: move before/after update hooks out of graphql directory
